device: Disable USB persist when suspending
USB persist means that the kernel will try to match up the new device
with the old device after resuming. This is useful in same cases, for
example for USB drivers where the file system can continue to use the
device.

However, the device will have been reset and will be in a bad state. As
such, we really don't want that to happen.

Note that we could probably just disable USB persist for all fingerprint
devices through the hwdb.